RDKEMW-1017: MiracastService COMRPC Support
