+++ Miracast/MiracastService/MiracastController.h
@@ -79,7 +79,7 @@ public:
 
     MiracastError stop_discover_devices(bool isNotificationRequired = true);
     MiracastError set_WFDParameters(void);
-    void restart_session_discovery(std::string& mac_address);
+    void restart_session_discovery(const std::string& mac_address);
     void flush_current_session(void);
     void remove_P2PGroupInstance(void);
     void restart_session(bool start_discovering_enabled);
