+++ Miracast/MiracastService/MiracastService.conf.in
@@ -1,4 +1,12 @@
 precondition = ["Platform"]
 callsign = "org.rdk.MiracastService"
-autostart = "false"
+autostart = "@PLUGIN_MIRACASTSERVICE_AUTOSTART@"
 startuporder = "@PLUGIN_MIRACASTSERVICE_STARTUPORDER@"
+
+configuration = JSON()
+rootobject = JSON()
+
+rootobject.add("mode", "@PLUGIN_MIRACASTSERVICE_MODE@")
+rootobject.add("locator", "lib@PLUGIN_IMPLEMENTATION@.so")
+
+configuration.add("root", rootobject)
