+++ Miracast/MiracastService/MiracastService.config
@@ -1,7 +1,16 @@
-set (autostart false)
+set (autostart ${PLUGIN_MIRACASTSERVICE_AUTOSTART})
 set (preconditions Platform)
 set (callsign "org.rdk.MiracastService")
 
 if(PLUGIN_MIRACASTSERVICE_STARTUPORDER)
 set (startuporder ${PLUGIN_MIRACASTSERVICE_STARTUPORDER})
 endif()
+
+map()
+    key(root)
+    map()
+        kv(mode ${PLUGIN_MIRACASTSERVICE_MODE})
+        kv(locator lib${PLUGIN_IMPLEMENTATION}.so)
+    end()
+end()
+ans(configuration)
